Refactor editor to own file
